+++ include/nodepp/utf.h
@@ -32,15 +32,15 @@ inline ptr_t<uint32> utf8_to_utf32( const ptr_t<uint8>& utf8 ) {
             codepoint = byte; i += 1;
         } elif ((byte & 0xE0) == 0xC0) {
             if (i + 1 >= utf8.size()){ throw except_t("Invalid UTF-8 sequence"); return nullptr; }
-            codepoint = (byte & 0x1F) << 6 | (utf8[i + 1] & 0x3F);
+            codepoint = (uint32)((byte & 0x1F) << 6 | (utf8[i + 1] & 0x3F));
             i += 2;
         } elif ((byte & 0xF0) == 0xE0) {
             if (i + 2 >= utf8.size()){ throw except_t("Invalid UTF-8 sequence"); return nullptr; }
-            codepoint = (byte & 0x0F) << 12 | (utf8[i + 1] & 0x3F) << 6 | (utf8[i + 2] & 0x3F);
+            codepoint = (uint32)((byte & 0x0F) << 12 | (utf8[i + 1] & 0x3F) << 6 | (utf8[i + 2] & 0x3F));
             i += 3;
         } elif ((byte & 0xF8) == 0xF0) {
             if (i + 3 >= utf8.size()){ throw except_t("Invalid UTF-8 sequence"); return nullptr; }
-            codepoint = (byte & 0x07) << 18 | (utf8[i + 1] & 0x3F) << 12 | (utf8[i + 2] & 0x3F) << 6 | (utf8[i + 3] & 0x3F);
+            codepoint = (uint32)((byte & 0x07) << 18 | (utf8[i + 1] & 0x3F) << 12 | (utf8[i + 2] & 0x3F) << 6 | (utf8[i + 3] & 0x3F));
             i += 4;
         } else { throw except_t("Invalid UTF-8 byte"); return nullptr; }
 
@@ -66,7 +66,7 @@ inline ptr_t<uint8> utf32_to_utf8( const ptr_t<uint32>& utf32 ) {
             utf8.push( type::cast<uint8>(((codepoint >> 12) & 0x3F) | 0x80) );
             utf8.push( type::cast<uint8>(((codepoint >> 6)  & 0x3F) | 0x80) );
             utf8.push( type::cast<uint8>((codepoint & 0x3F) | 0x80) );
-        } else { 
+        } else {
             throw except_t("Invalid UTF-32 codepoint"); return nullptr;
         }
     }   return utf8.data();
@@ -82,15 +82,15 @@ inline ptr_t<uint16> utf8_to_utf16( const ptr_t<uint8>& utf8 ) {
             codepoint = byte; i += 1;
         } elif ((byte & 0xE0) == 0xC0) {
             if (i+1 >= utf8.size()){ throw except_t("Invalid UTF-8 sequence"); return nullptr; }
-            codepoint = (byte & 0x1F) << 6 | (utf8[i + 1] & 0x3F); 
+            codepoint = (uint32)((byte & 0x1F) << 6 | (utf8[i + 1] & 0x3F));
             i += 2;
         } elif ((byte & 0xF0) == 0xE0) {
             if (i+2 >= utf8.size()){ throw except_t("Invalid UTF-8 sequence"); return nullptr; }
-            codepoint = (byte & 0x0F) << 12 | (utf8[i + 1] & 0x3F) << 6 | (utf8[i + 2] & 0x3F);
+            codepoint = (uint32)((byte & 0x0F) << 12 | (utf8[i + 1] & 0x3F) << 6 | (utf8[i + 2] & 0x3F));
             i += 3;
         } elif ((byte & 0xF8) == 0xF0) {
             if (i+3 >= utf8.size()){ throw except_t("Invalid UTF-8 sequence"); return nullptr; }
-            codepoint = (byte & 0x07) << 18 | (utf8[i + 1] & 0x3F) << 12 | (utf8[i + 2] & 0x3F) << 6 | (utf8[i + 3] & 0x3F);
+            codepoint = (uint32)((byte & 0x07) << 18 | (utf8[i + 1] & 0x3F) << 12 | (utf8[i + 2] & 0x3F) << 6 | (utf8[i + 3] & 0x3F));
             i += 4;
         } else { throw except_t("Invalid UTF-8 sequence"); return nullptr; }
 
@@ -111,7 +111,7 @@ inline ptr_t<uint8> utf16_to_utf8(const ptr_t<uint16>& utf16) {
     queue_t<uint8> utf8; ulong i = 0; while ( i<utf16.size() ) {
         uint32 codepoint = 0; uint16 unit = utf16[i];
 
-        if (unit < 0xD800 || unit > 0xDFFF) { 
+        if (unit < 0xD800 || unit > 0xDFFF) {
             codepoint = unit; i += 1;
         } elif (unit >= 0xD800 && unit <= 0xDBFF) {
             if (i + 1 >= utf16.size()){ throw except_t("Invalid UTF-16 sequence"); return nullptr; }
@@ -155,7 +155,7 @@ inline ptr_t<uint32> utf16_to_utf32( const ptr_t<uint16>& utf16 ) {
             uint32 codepoint = ((unit-0xD800)<<10) + (low-0xDC00) + 0x10000;
             utf32.push( codepoint ); i += 2;
         } else { throw except_t("Invalid UTF-16 high surrogate"); return nullptr; }
-    
+
     }   return utf32.data();
 }
